cdk-workshop/pyproject.toml
[project]
name = "cdk-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.201.0",
    "boto3>=1.39.2",
    "constructs>=10.0.0,<11.0.0",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
